Split Ironfishmanger types into files.
