+++ electron/dev/index.ts
@@ -2,11 +2,9 @@
 /* eslint-disable @typescript-eslint/ban-ts-comment */
 //@ts-nocheck
 import { ipcMain } from 'electron'
-import {
-  IronfishAccountManagerAction,
-  IronfishManagerAction,
-  IronfishTransactionManagerAction,
-} from 'Types/IIronfishManager'
+import { IronfishManagerAction } from 'Types/IronfishManager/IIronfishManager'
+import { IronfishTransactionManagerAction } from 'Types/IronfishManager/IIronfishTransactionManager'
+import { IronfishAccountManagerAction } from 'Types/IronfishManager/IIronfishAccountManager'
 import initStorageCallbacks from '../common/initStorage'
 import { IronFishManager } from '../ironfish/IronFishManager'
 import '../common/index'
