+++ src/initGlobalVariables.ts
@@ -2,7 +2,7 @@ import DemoDataManager from './data/DemoDataManager'
 import IStorage from 'Types/IStorage'
 import Contact from 'Types/Contact'
 import AccountSettings from 'Types/AccountSettings'
-import IIronfishManager from 'Types/IIronfishManager'
+import IIronfishManager from 'Types/IronfishManager/IIronfishManager'
 
 declare global {
   interface Window {
