+++ src/withDemoData.ts
@@ -2,7 +2,8 @@ import { AccountValue } from '@ironfish/sdk'
 import noop from 'lodash/noop'
 import AccountSettings from 'Types/AccountSettings'
 import Contact from 'Types/Contact'
-import IIronfishManager, { TransactionReceiver } from 'Types/IIronfishManager'
+import IIronfishManager from 'Types/IronfishManager/IIronfishManager'
+import { TransactionReceiver } from 'Types/IronfishManager/IIronfishTransactionManager'
 import IStorage from 'Types/IStorage'
 import SortType from 'Types/SortType'
