+++ types/IronfishManager/IIronfishManager.ts
@@ -1,0 +1,32 @@
+import { PeerResponse } from '@ironfish/sdk'
+import IronFishInitStatus from '../IronfishInitStatus'
+import NodeStatusResponse from 'Types/NodeStatusResponse'
+
+import { IIronfishAccountManager } from './IIronfishAccountManager'
+import { IIronfishTransactionManager } from './IIronfishTransactionManager'
+
+export enum IronfishManagerAction {
+  INITIALIZE = 'initialize',
+  START = 'start',
+  STOP = 'stop',
+  STATUS = 'status',
+  NODE_STATUS = 'nodeStatus',
+  PEERS = 'peers',
+  HAS_ANY_ACCOUNT = 'hasAnyAccount',
+  SYNC = 'sync',
+}
+
+export interface IIronfishManager {
+  accounts: IIronfishAccountManager
+  transactions: IIronfishTransactionManager
+  initialize: () => Promise<void>
+  hasAnyAccount: () => Promise<boolean>
+  start: () => Promise<void>
+  stop: () => Promise<void>
+  status: () => Promise<IronFishInitStatus>
+  sync: () => Promise<void>
+  nodeStatus: () => Promise<NodeStatusResponse>
+  peers: () => Promise<PeerResponse[]>
+}
+
+export default IIronfishManager
